+++ training/src/pages/Trainee/components/index.js
@@ -1,1 +1,1 @@
-export { AddDialog } from './AddDialog';
+export { AddDialog } from './AddDialog/index';
